# PERF: use python int in RangeIndex.get_loc